Add creation of states based on maze size

+++ program/Main/include/Arguments.hpp
@@ -12,7 +12,7 @@ using namespace std;
 
 class Arguments {
 private: 
-    int puzzleSize;
+    uint8_t puzzleSize;
     std::vector<char> firstMoves;
     State firstStateFromFile;
     AbstractAlgorithmPointer algorithm;
